Integrated book classes into recommendation.java.
This integrates the books classes that I worked on with the
recommendation.java file created by Chris B. I have also made it so that
it will compile and run with tests made in each main methods for book,
bookrating, and recommendation. You will notice the import statements at
the top of each java file. The user classes have not been integrated.